tests/requirements.txt
python-binance==1.0.28
requests==2.32.3
